+++ jkube-kit/common/src/test/java/org/eclipse/jkube/kit/common/archive/AssemblyFileSetUtilsTest.java
@@ -14,9 +14,10 @@
 package org.eclipse.jkube.kit.common.archive;
 
 import java.io.File;
-import java.util.Map;
+import java.util.List;
 
 import org.apache.commons.io.FileUtils;
+import org.eclipse.jkube.kit.common.AssemblyFileEntry;
 import org.eclipse.jkube.kit.common.AssemblyFileSet;
 
 import org.junit.Rule;
@@ -25,9 +26,10 @@ import org.junit.rules.TemporaryFolder;
 
 import static org.eclipse.jkube.kit.common.archive.AssemblyFileSetUtils.calculateFilePermissions;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.contains;
+import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.hasEntry;
 import static org.hamcrest.Matchers.hasSize;
 
 public class AssemblyFileSetUtilsTest {
@@ -39,69 +41,86 @@ public class AssemblyFileSetUtilsTest {
   public void calculateFilePermissionsFileWithNoFileMode() throws Exception {
     // Given
     final AssemblyFileSet afs = AssemblyFileSet.builder().build();
+    final File sourceFile = temp.newFile("source-file.txt");
     final File aFile = temp.newFile("just-a-file.txt");
     // When
-    final Map<File, String> result = calculateFilePermissions(aFile, afs);
+    final List<AssemblyFileEntry> result = calculateFilePermissions(sourceFile, aFile, afs);
     // Then
-    assertThat(result.entrySet(), empty());
+    assertThat(result, empty());
   }
 
   @Test
   public void calculateFilePermissionsFileWithFileMode() throws Exception {
     // Given
     final AssemblyFileSet afs = AssemblyFileSet.builder().fileMode("0644").build();
+    final File sourceFile = temp.newFile("source-file.txt");
     final File aFile = temp.newFile("just-a-file.txt");
     // When
-    final Map<File, String> result = calculateFilePermissions(aFile, afs);
+    final List<AssemblyFileEntry> result = calculateFilePermissions(sourceFile, aFile, afs);
     // Then
-    assertThat(result.entrySet(), hasSize(1));
-    assertThat(result, hasEntry(aFile, "0644"));
+    assertThat(result, hasSize(1));
+    assertThat(result, contains(new AssemblyFileEntry(sourceFile, aFile, "0644")));
   }
 
   @Test
   public void calculateFilePermissionsDirectoryWithNoDirectoryMode() throws Exception {
     // Given
     final AssemblyFileSet afs = AssemblyFileSet.builder().build();
+    final File sourceDirectory = temp.newFile("source-directory");
     final File aDirectory = temp.newFolder("just-a-directory");
     // When
-    final Map<File, String> result = calculateFilePermissions(aDirectory, afs);
+    final List<AssemblyFileEntry> result = calculateFilePermissions(sourceDirectory, aDirectory, afs);
     // Then
-    assertThat(result.entrySet(), hasSize(1));
-    assertThat(result, hasEntry(aDirectory, "040111"));
+    assertThat(result, hasSize(1));
+    assertThat(result, contains(new AssemblyFileEntry(sourceDirectory, aDirectory, "040111")));
   }
 
   @Test
   public void calculateFilePermissionsDirectoryWithDirectoryMode() throws Exception {
     // Given
     final AssemblyFileSet afs = AssemblyFileSet.builder().directoryMode("040755").build();
+    final File sourceDirectory = temp.newFolder("source-directory");
+    final File sourceSubdirectory = new File(sourceDirectory, "subdirectory");
+    FileUtils.forceMkdir(sourceSubdirectory);
+    final File sourceFile = new File(sourceDirectory, "file.txt");
+    assertThat(sourceFile.createNewFile(), equalTo(true));
     final File aDirectory = temp.newFolder("just-a-directory");
     final File aSubdirectory = new File(aDirectory, "subdirectory");
     FileUtils.forceMkdir(aSubdirectory);
     final File aFile = new File(aDirectory, "file.txt");
     assertThat(aFile.createNewFile(), equalTo(true));
     // When
-    final Map<File, String> result = calculateFilePermissions(aDirectory, afs);
+    final List<AssemblyFileEntry> result = calculateFilePermissions(sourceDirectory, aDirectory, afs);
     // Then
-    assertThat(result.entrySet(), hasSize(2));
-    assertThat(result, hasEntry(aDirectory, "040755"));
-    assertThat(result, hasEntry(aSubdirectory, "040755"));
+    assertThat(result, hasSize(2));
+    assertThat(result, containsInAnyOrder(
+        new AssemblyFileEntry(sourceDirectory, aDirectory, "040755"),
+        new AssemblyFileEntry(sourceSubdirectory, aSubdirectory, "040755")
+    ));
   }
 
   @Test
   public void calculateFilePermissionsDirectoryWithDirectoryAndFileMode() throws Exception {
     // Given
     final AssemblyFileSet afs = AssemblyFileSet.builder().directoryMode("040755").fileMode("0644").build();
+    final File sourceDirectory = temp.newFolder("source-directory");
+    final File sourceSubdirectory = new File(sourceDirectory, "subdirectory");
+    FileUtils.forceMkdir(sourceSubdirectory);
+    final File sourceFile = new File(sourceDirectory, "file.txt");
+    assertThat(sourceFile.createNewFile(), equalTo(true));
     final File aDirectory = temp.newFolder("just-a-directory");
     final File aSubdirectory = new File(aDirectory, "subdirectory");
     FileUtils.forceMkdir(aSubdirectory);
     final File aFile = new File(aDirectory, "file.txt");
     assertThat(aFile.createNewFile(), equalTo(true));
     // When
-    final Map<File, String> result = calculateFilePermissions(aDirectory, afs);
+    final List<AssemblyFileEntry> result = calculateFilePermissions(sourceDirectory, aDirectory, afs);
     // Then
-    assertThat(result.entrySet(), hasSize(3));
-    assertThat(result, hasEntry(aDirectory, "040755"));
-    assertThat(result, hasEntry(aSubdirectory, "040755"));
-    assertThat(result, hasEntry(aFile, "0644"));
+    assertThat(result, hasSize(3));
+    assertThat(result, containsInAnyOrder(
+        new AssemblyFileEntry(sourceDirectory, aDirectory, "040755"),
+        new AssemblyFileEntry(sourceSubdirectory, aSubdirectory, "040755"),
+        new AssemblyFileEntry(sourceFile, aFile, "0644")
+    ));
   }
 }