+++ jkube-kit/watcher/api/src/main/java/org/eclipse/jkube/watcher/api/WatcherContext.java
@@ -13,22 +13,18 @@
  */
 package org.eclipse.jkube.watcher.api;
 
-import io.fabric8.kubernetes.client.KubernetesClient;
+import org.eclipse.jkube.kit.build.service.docker.WatchService;
+import org.eclipse.jkube.kit.common.KitLogger;
+import org.eclipse.jkube.kit.config.JKubeConfiguration;
+import org.eclipse.jkube.kit.config.resource.ProcessorConfig;
+import org.eclipse.jkube.kit.config.service.JKubeServiceHub;
+
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.eclipse.jkube.kit.config.JKubeConfiguration;
-import org.eclipse.jkube.kit.common.JavaProject;
-import org.eclipse.jkube.kit.build.service.docker.ServiceHub;
-import org.eclipse.jkube.kit.build.service.docker.WatchService;
-import org.eclipse.jkube.kit.common.KitLogger;
-import org.eclipse.jkube.kit.config.access.ClusterConfiguration;
-import org.eclipse.jkube.kit.config.resource.RuntimeMode;
-import org.eclipse.jkube.kit.config.resource.ProcessorConfig;
-import org.eclipse.jkube.kit.config.service.JKubeServiceHub;
 
 /**
  * @author nicola
@@ -40,18 +36,13 @@ import org.eclipse.jkube.kit.config.service.JKubeServiceHub;
 @EqualsAndHashCode
 public class WatcherContext {
 
-    private JavaProject project;
-    private ProcessorConfig config;
-    private KitLogger logger;
-    private KitLogger newPodLogger;
-    private KitLogger oldPodLogger;
-    private RuntimeMode mode;
-    private boolean useProjectClasspath;
-    private ServiceHub serviceHub;
-    private WatchService.WatchContext watchContext;
-    private JKubeConfiguration buildContext;
-    private ClusterConfiguration clusterConfiguration;
-    private KubernetesClient kubernetesClient;
-    private JKubeServiceHub jKubeServiceHub;
+  private ProcessorConfig config;
+  private KitLogger logger;
+  private KitLogger newPodLogger;
+  private KitLogger oldPodLogger;
+  private boolean useProjectClasspath;
+  private WatchService.WatchContext watchContext;
+  private JKubeConfiguration buildContext;
+  private JKubeServiceHub jKubeServiceHub;
 
 }